update date 22/09/2023

+++ grocery_store_sales_backend/grocery_store_sales_backend/src/main/java/com/example/grocery_store_sales_backend/repository/IUserRepository.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IUserRepository extends JpaRepository<Users,Long> {
+public interface IUserRepository extends JpaRepository<Users,Integer> {
     Optional<Users> findByEmailUser(String email);
+
 }
